restructure; add to kex map
