--- russh/src/kex/ecdh/mod.rs
@@ -1,1 +1,0 @@
-mod nistp256;
