+++ russh/src/kex/mod.rs
@@ -17,7 +17,7 @@
 //! This module exports kex algorithm names for use with [Preferred].
 mod curve25519;
 mod dh;
-mod ecdh;
+mod ecdh_nistp;
 mod none;
 use std::cell::RefCell;
 use std::collections::HashMap;
@@ -28,6 +28,7 @@ use dh::{
     DhGroup14Sha1KexType, DhGroup14Sha256KexType, DhGroup16Sha512KexType, DhGroup1Sha1KexType,
 };
 use digest::Digest;
+use ecdh_nistp::{EcdhNistP256KexType, EcdhNistP384KexType, EcdhNistP521KexType};
 use once_cell::sync::Lazy;
 use russh_cryptovec::CryptoVec;
 use russh_keys::encoding::Encoding;
@@ -98,6 +99,12 @@ pub const DH_G14_SHA1: Name = Name("diffie-hellman-group14-sha1");
 pub const DH_G14_SHA256: Name = Name("diffie-hellman-group14-sha256");
 /// `diffie-hellman-group16-sha512`
 pub const DH_G16_SHA512: Name = Name("diffie-hellman-group16-sha512");
+/// `ecdh-sha2-nistp256`
+pub const ECDH_SHA2_NISTP256: Name = Name("ecdh-sha2-nistp256");
+/// `ecdh-sha2-nistp384`
+pub const ECDH_SHA2_NISTP384: Name = Name("ecdh-sha2-nistp384");
+/// `ecdh-sha2-nistp521`
+pub const ECDH_SHA2_NISTP521: Name = Name("ecdh-sha2-nistp521");
 /// `none`
 pub const NONE: Name = Name("none");
 /// `ext-info-c`
@@ -114,6 +121,9 @@ const _DH_G1_SHA1: DhGroup1Sha1KexType = DhGroup1Sha1KexType {};
 const _DH_G14_SHA1: DhGroup14Sha1KexType = DhGroup14Sha1KexType {};
 const _DH_G14_SHA256: DhGroup14Sha256KexType = DhGroup14Sha256KexType {};
 const _DH_G16_SHA512: DhGroup16Sha512KexType = DhGroup16Sha512KexType {};
+const _ECDH_SHA2_NISTP256: EcdhNistP256KexType = EcdhNistP256KexType {};
+const _ECDH_SHA2_NISTP384: EcdhNistP384KexType = EcdhNistP384KexType {};
+const _ECDH_SHA2_NISTP521: EcdhNistP521KexType = EcdhNistP521KexType {};
 const _NONE: none::NoneKexType = none::NoneKexType {};
 
 pub(crate) static KEXES: Lazy<HashMap<&'static Name, &(dyn KexType + Send + Sync)>> =
@@ -125,6 +135,9 @@ pub(crate) static KEXES: Lazy<HashMap<&'static Name, &(dyn KexType + Send + Sync
         h.insert(&DH_G14_SHA256, &_DH_G14_SHA256);
         h.insert(&DH_G14_SHA1, &_DH_G14_SHA1);
         h.insert(&DH_G1_SHA1, &_DH_G1_SHA1);
+        h.insert(&ECDH_SHA2_NISTP256, &_ECDH_SHA2_NISTP256);
+        h.insert(&ECDH_SHA2_NISTP384, &_ECDH_SHA2_NISTP384);
+        h.insert(&ECDH_SHA2_NISTP521, &_ECDH_SHA2_NISTP521);
         h.insert(&NONE, &_NONE);
         h
     });
